feat: refactor a class context
